pyproject.toml
[project]
name = "ai-hackaton"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10, <3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "streamlit>=1.55.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "ultralytics>=8.4.26",
    "pydantic>=2.12.5",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"
